+++ src/Files.Backend/Services/IPreviewPopupProvider.cs
@@ -1,0 +1,30 @@
+﻿// Copyright (c) 2023 Files Community
+// Licensed under the MIT License. See the LICENSE.
+
+using System.Threading.Tasks;
+
+namespace Files.Backend.Services
+{
+	public interface IPreviewPopupProvider
+	{
+		/// <summary>
+		/// Toggle preview popup
+		/// </summary>
+		/// <param name="path"></param>
+		/// <returns></returns>
+		Task TogglePreviewPopup(string path);
+
+		/// <summary>
+		/// Switch preview
+		/// </summary>
+		/// <param name="path"></param>
+		/// <returns></returns>
+		Task SwitchPreview(string path);
+
+		/// <summary>
+		/// Detect if the provider is available
+		/// </summary>
+		/// <returns></returns>
+		Task<bool> DetectAvailability();
+	}
+}
